+++ src/android/io/branch/BranchSDK.java
@@ -22,13 +22,10 @@ import java.io.IOException;
 import io.branch.indexing.BranchUniversalObject;
 import io.branch.referral.Branch;
 import io.branch.referral.BranchError;
-import io.branch.referral.BranchViewHandler;
-import io.branch.referral.ServerRequestGetCPID.BranchCrossPlatformIdListener;
 import io.branch.referral.ServerRequestGetLATD.BranchLastAttributedTouchDataListener;
 import io.branch.referral.SharingHelper;
 import io.branch.referral.QRCode.BranchQRCode;
 import io.branch.referral.util.BRANCH_STANDARD_EVENT;
-import io.branch.referral.util.BranchCPID;
 import io.branch.referral.util.BranchEvent;
 import io.branch.referral.util.ContentMetadata;
 import io.branch.referral.util.CurrencyType;
@@ -42,7 +39,6 @@ public class BranchSDK extends CordovaPlugin {
 
     // Standard Debugging Variables
     private static final String LCAT = "CordovaBranchSDK";
-
     private static final String BRANCH_PLUGIN_TYPE = "CordovaIonic";
     private static final String BRANCH_PLUGIN_VERSION = "%BRANCH_PLUGIN_VERSION%";
 
@@ -86,23 +82,6 @@ public class BranchSDK extends CordovaPlugin {
     }
 
     /**
-     * Handle depreciated call to sendJavaScript for more recent method
-     */
-    @TargetApi(Build.VERSION_CODES.KITKAT)
-    private void sendJavascript(final String javascript) {
-        webView.getView().post(new Runnable() {
-            @Override
-            public void run() {
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-                    webView.sendJavascript(javascript);
-                } else {
-                    webView.loadUrl("javascript:" + javascript);
-                }
-            }
-        });
-    }
-
-    /**
      * <p>
      * cordova.exec() method reference.
      * All exec() calls goes to this part.
@@ -122,9 +101,6 @@ public class BranchSDK extends CordovaPlugin {
         if (action.equals("enableLogging")) {
             cordova.getActivity().runOnUiThread(r);
             return true;
-        } else if (action.equals("setCookieBasedMatching")) {
-            cordova.getActivity().runOnUiThread(r);
-            return true;
        } else if (action.equals("disableTracking")) {
             cordova.getActivity().runOnUiThread(r);
             return true;
@@ -139,16 +115,9 @@ public class BranchSDK extends CordovaPlugin {
                 if (action.equals("setIdentity")) {
                     cordova.getActivity().runOnUiThread(r);
                     return true;
-                } else if (action.equals("userCompletedAction")) {
-                    if (args.length() < 1 && args.length() > 2) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
-                        return false;
-                    }
-                    cordova.getActivity().runOnUiThread(r);
-                    return true;
                 } else if (action.equals("sendBranchEvent")) {
                     if (args.length() < 1 && args.length() > 2) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
+                        callbackContext.error("Parameter count mismatch");
                         return false;
                     }
                     cordova.getActivity().runOnUiThread(r);
@@ -164,37 +133,33 @@ public class BranchSDK extends CordovaPlugin {
                     return true;
                 } else if (action.equals("createBranchUniversalObject")) {
                     if (args.length() != 1) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
+                        callbackContext.error("Parameter count mismatch");
                         return false;
                     }
                     cordova.getActivity().runOnUiThread(r);
                     return true;
 
-                } else if (action.equals("crossPlatformIds")) {
+                } else if (action.equals("lastAttributedTouchData")) {
                     cordova.getActivity().runOnUiThread(r);
                     return true;
 
-                } else if (action.equals("lastAttributedTouchData")) {
-                    cordova.getActivity().runOnUiThread(r);
-                    return true; 
-
                 } else if (action.equals(("generateShortUrl"))) {
                     if (args.length() != 3) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
+                        callbackContext.error("Parameter count mismatch");
                         return false;
                     }
                     cordova.getActivity().runOnUiThread(r);
                     return true;
                 } else if (action.equals("registerView")) {
                     if (args.length() != 1) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
+                        callbackContext.error("Parameter count mismatch");
                         return false;
                     }
                     cordova.getActivity().runOnUiThread(r);
                     return true;
                 } else if (action.equals("showShareSheet")) {
                     if (args.length() < 3) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
+                        callbackContext.error("Parameter count mismatch");
                         return false;
                     }
                     cordova.getActivity().runOnUiThread(r);
@@ -229,11 +194,11 @@ public class BranchSDK extends CordovaPlugin {
 
                 } else if (action.equals("getBranchQRCode")) {
                       if (args.length() != 4) {
-                        callbackContext.error(String.format("Parameter count mismatch"));
+                        callbackContext.error("Parameter count mismatch");
                         return false;
                     }
                     cordova.getActivity().runOnUiThread(r);
-                    return true; 
+                    return true;
                 }
 
                 return true;
@@ -245,10 +210,6 @@ public class BranchSDK extends CordovaPlugin {
 
         return false;
 
-    }
-    
-    public void crossPlatformIds(CallbackContext callbackContext) {
-        this.instance.getCrossPlatformIds(new BranchCPIDListener(callbackContext));
     }
 
     public void lastAttributedTouchData(CallbackContext callbackContext) {
@@ -275,7 +236,7 @@ public class BranchSDK extends CordovaPlugin {
             this.deepLinkUrl = data.toString();
         }
 
-        this.instance.initSession(new SessionListener(callbackContext), data, activity);
+        Branch.sessionBuilder(activity).withData(data).withCallback(new SessionListener(callbackContext)).init();
     }
 
     /**
@@ -584,24 +545,6 @@ public class BranchSDK extends CordovaPlugin {
     }
 
     /**
-     * <p>Sets the cookie based matching for all incoming requests.</p>
-     * <p>If you want cookie based matching, call this <b>before</b> initUserSession</p>
-     *
-     * @param linkDomain      A {@link String} value to of the link domain for cookie based matching.
-     * @param callbackContext A callback to execute at the end of this method
-     */
-    private void setCookieBasedMatching(String linkDomain, CallbackContext callbackContext) {
-
-        this.activity = this.cordova.getActivity();
-
-        if (linkDomain != null) {
-            Branch.enableCookieBasedMatching(linkDomain);
-        }
-
-        callbackContext.success("Success");
-    }
-
-    /**
      * <p>Enabling Branch SDK logging</p>
      *
      * @param isEnable        A {@link Boolean} value to enable/disable logging
@@ -649,46 +592,13 @@ public class BranchSDK extends CordovaPlugin {
     /**
      * <p>Allow Branch SDK to pass the user's Mixpanel distinct id to our servers. Branch will then pass that Distinct ID to Mixpanel when logging any event.</p>
      *
-     * @param token           A {@link String} value containing the unique identifier of the Mixpanel user.
+     * @param key           A {@link String} value containing the key for the unique identifier of the Mixpanel user.
+     * @param val           A {@link String} value containing the value for the unique identifier of the Mixpanel user.
      * @param callbackContext A callback to execute at the end of this method
      */
     private void setRequestMetadata(String key, String val, CallbackContext callbackContext) {
 
         Branch.getInstance().setRequestMetadata(key, val);
-
-        callbackContext.success("Success");
-
-    }
-
-    /**
-     * <p>A void call to indicate that the user has performed a specific action and for that to be
-     * reported to the Branch API.</p>
-     *
-     * @param action          A {@link String} value to be passed as an action that the user has carried out.
-     *                        For example "logged in" or "registered".
-     * @param callbackContext A callback to execute at the end of this method
-     */
-    private void userCompletedAction(String action, CallbackContext callbackContext) {
-
-        this.instance.userCompletedAction(action);
-
-        callbackContext.success("Success");
-
-    }
-
-    /**
-     * <p>A void call to indicate that the user has performed a specific action and for that to be
-     * reported to the Branch API.</p>
-     *
-     * @param action          A {@link String} value to be passed as an action that the user has carried
-     *                        out. For example "logged in" or "registered".
-     * @param metaData        A {@link JSONObject} containing app-defined meta-data to be attached to a
-     *                        user action that has just been completed.
-     * @param callbackContext A callback to execute at the end of this method
-     */
-    private void userCompletedAction(String action, JSONObject metaData, CallbackContext callbackContext) {
-
-        this.instance.userCompletedAction(action, metaData);
 
         callbackContext.success("Success");
 
@@ -841,39 +751,6 @@ public class BranchSDK extends CordovaPlugin {
     //----------- INNER CLASS LISTENERS ------------//
     //////////////////////////////////////////////////
 
-    protected class BranchCPIDListener implements BranchCrossPlatformIdListener {
-        private CallbackContext _callbackContext;
-
-        public BranchCPIDListener(CallbackContext callbackContext) {
-            this._callbackContext = callbackContext;
-        }
-
-        @Override
-        public void onDataFetched(BranchCPID branchCPID, BranchError error) {
-            if (error != null) {
-                Log.d(LCAT, "CPID unavailable");
-                this._callbackContext.error("CPID unavailable");
-            } else {
-
-                JSONObject jsonObject = new JSONObject();
-                try {
-                    jsonObject.put("developer_identity", branchCPID.getDeveloperIdentity());
-                    jsonObject.put("cross_platform_id", branchCPID.getCrossPlatformID());
-                    jsonObject.put("past_cross_platform_ids", branchCPID.getPastCrossPlatformIds());
-                    jsonObject.put("prob_cross_platform_ids", branchCPID.getProbabilisticCrossPlatformIds());
-                } catch (JSONException e) {
-                    // just send back and empty object on json error
-                    jsonObject = new JSONObject();
-                }
-
-                Log.d(LCAT, jsonObject.toString());
-
-                PluginResult result = new PluginResult(PluginResult.Status.OK, jsonObject);
-                this._callbackContext.sendPluginResult(result);
-            }
-        }
-    }
-
     protected class BranchLATDListener implements BranchLastAttributedTouchDataListener {
         private CallbackContext _callbackContext;
 
@@ -892,35 +769,6 @@ public class BranchSDK extends CordovaPlugin {
                 PluginResult result = new PluginResult(PluginResult.Status.OK, jsonObject);
                 this._callbackContext.sendPluginResult(result);
             }
-        }
-    }
-
-    protected class BranchViewEventsListener implements BranchViewHandler.IBranchViewEvents {
-
-        private CallbackContext _callbackContext;
-
-        public BranchViewEventsListener(CallbackContext callbackContext) {
-            callbackContext.success("Success");
-        }
-
-        @Override
-        public void onBranchViewVisible(String action, String branchViewID) {
-
-        }
-
-        @Override
-        public void onBranchViewAccepted(String action, String branchViewID) {
-
-        }
-
-        @Override
-        public void onBranchViewCancelled(String action, String branchViewID) {
-
-        }
-
-        @Override
-        public void onBranchViewError(int errorCode, String errorMsg, String action) {
-
         }
     }
 
@@ -1235,8 +1083,6 @@ public class BranchSDK extends CordovaPlugin {
 
                 if (this.action.equals("enableLogging")) {
                     enableLogging(this.args.getBoolean(0), this.callbackContext);
-                } else if (this.action.equals("setCookieBasedMatching")) {
-                    setCookieBasedMatching(this.args.getString(0), this.callbackContext);
                 } else if (this.action.equals("disableTracking")) {
                     disableTracking(this.args.getBoolean(0), this.callbackContext);
                 } else if (this.action.equals("initSession")) {
@@ -1246,12 +1092,6 @@ public class BranchSDK extends CordovaPlugin {
                 } else {
                     if (this.action.equals("setIdentity")) {
                         setIdentity(this.args.getString(0), this.callbackContext);
-                    } else if (this.action.equals("userCompletedAction")) {
-                        if (this.args.length() == 2) {
-                            userCompletedAction(this.args.getString(0), this.args.getJSONObject(1), this.callbackContext);
-                        } else if (this.args.length() == 1) {
-                            userCompletedAction(this.args.getString(0), this.callbackContext);
-                        }
                     } else if (this.action.equals("sendBranchEvent")) {
                         if (this.args.length() == 2) {
                             sendBranchEvent(this.args.getString(0), this.args.getJSONObject(1), this.callbackContext);
@@ -1266,8 +1106,6 @@ public class BranchSDK extends CordovaPlugin {
                         logout(this.callbackContext);
                     } else if (this.action.equals("createBranchUniversalObject")) {
                         createBranchUniversalObject(this.args.getJSONObject(0), this.callbackContext);
-                    } else if (this.action.equals("crossPlatformIds")) {
-                        crossPlatformIds(this.callbackContext);
                     } else if (this.action.equals("lastAttributedTouchData")) {
                         lastAttributedTouchData(this.callbackContext);
                     } else if (this.action.equals(("generateShortUrl"))) {