+++ src/ios/BranchSDK.h
@@ -2,12 +2,16 @@
 
 #ifdef BRANCH_NPM
 #import "Branch.h"
+#import "BranchEvent.h"
+#import "BranchQRCode.h"
 #import "BranchLinkProperties.h"
 #import "BranchUniversalObject.h"
 #else
-#import <Branch/Branch.h>
-#import <Branch/BranchLinkProperties.h>
-#import <Branch/BranchUniversalObject.h>
+#import <BranchSDK/Branch.h>
+#import <BranchSDK/BranchEvent.h>
+#import <BranchSDK/BranchQRCode.h>
+#import <BranchSDK/BranchLinkProperties.h>
+#import <BranchSDK/BranchUniversalObject.h>
 #endif
 
 #import <Cordova/CDV.h>
@@ -35,9 +39,7 @@
 - (void)getFirstReferringParams:(CDVInvokedUrlCommand*)command;
 - (void)setIdentity:(CDVInvokedUrlCommand*)command;
 - (void)registerDeepLinkController:(CDVInvokedUrlCommand*)command;
-- (void)userCompletedAction:(CDVInvokedUrlCommand*)command;
 - (void)logout:(CDVInvokedUrlCommand*)command;
-- (void)delayInitToCheckForSearchAds:(CDVInvokedUrlCommand*)command;
 
 // Branch Universal Object Methods
 - (void)createBranchUniversalObject:(CDVInvokedUrlCommand*)command;
@@ -49,7 +51,6 @@
 - (void)listOnSpotlight:(CDVInvokedUrlCommand*)command;
 
 // Branch Query Methods
-- (void)crossPlatformIds:(CDVInvokedUrlCommand *)command;
 - (void)lastAttributedTouchData:(CDVInvokedUrlCommand *)command;
 
 @end