+++ src/ios/BranchSDK.m
@@ -225,34 +225,6 @@ NSString * const pluginVersion = @"%BRANCH_PLUGIN_VERSION%";
   [branch registerDeepLinkController:controller forKey:[command.arguments objectAtIndex:0]];
 }
 
-- (void)userCompletedAction:(CDVInvokedUrlCommand*)command
-{
-  NSString *name;
-  NSDictionary *state;
-
-  // if a state dictionary is passed as an argument
-  if ([command.arguments count] == 2) {
-    name = [command.arguments objectAtIndex:0];
-    state = [command.arguments objectAtIndex:1];
-  }
-  else {
-    name = [command.arguments objectAtIndex:0];
-  }
-
-  Branch *branch = [self getInstance];
-
-  if (state) {
-    [branch userCompletedAction:name withState:state];
-  }
-  else {
-    [branch userCompletedAction:name];
-  }
-
-  // TODO: iOS Branch.userCompletedAction needs a callback for success or failure
-  CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString: @"Success"];
-  [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-}
-
 -(void)sendBranchEvent:(CDVInvokedUrlCommand*)command
 {
     NSString *eventName = [command.arguments objectAtIndex:0];
@@ -330,18 +302,6 @@ NSString * const pluginVersion = @"%BRANCH_PLUGIN_VERSION%";
   self.branchUniversalObjArray = [[NSMutableArray alloc] init];
 }
 
-- (void)delayInitToCheckForSearchAds:(CDVInvokedUrlCommand*)command
-{
-  bool enabled = [[command.arguments objectAtIndex:0] boolValue];
-  if (enabled) {
-    [[Branch getInstance] delayInitToCheckForSearchAds];
-  }
-
-  CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsBool:enabled];
-
-  [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-}
-
 #pragma mark - Branch Universal Object Methods
 
 - (void)createBranchUniversalObject:(CDVInvokedUrlCommand*)command
@@ -351,20 +311,17 @@ NSString * const pluginVersion = @"%BRANCH_PLUGIN_VERSION%";
 
   for (id key in properties) {
     if ([key isEqualToString:@"contentMetadata"]){
-      NSDictionary *metadata = (NSDictionary *)[properties valueForKey:key];
-
-      for (id key_ in metadata) {
-        [branchUniversalObj addMetadataKey:key_ value:[metadata valueForKey:key_]];
-      }
+        NSMutableDictionary<NSString *,NSString *> *metadata = (NSMutableDictionary<NSString *,NSString *> *)[properties valueForKey:key];
+        [[branchUniversalObj contentMetadata] setCustomMetadata:metadata];
     }
     else if ([key isEqualToString:@"contentIndexingMode"]) {
       NSString *indexingMode = [properties valueForKey:key];
       // Default contentIndexMode is always public
       if ([indexingMode isEqualToString:@"private"]) {
-        branchUniversalObj.contentIndexMode = BranchContentIndexModePrivate;
+        branchUniversalObj.publiclyIndex = false;
       }
       else {
-        branchUniversalObj.contentIndexMode = BranchContentIndexModePublic;
+        branchUniversalObj.publiclyIndex = true;
       }
     }
     else if ([key isEqualToString:@"canonicalIdentifier"]) {
@@ -598,37 +555,6 @@ NSString * const pluginVersion = @"%BRANCH_PLUGIN_VERSION%";
 }
 
 #pragma mark Branch Query Methods
-
-- (void)crossPlatformIds:(CDVInvokedUrlCommand *)command {
-  NSMutableDictionary *json = [NSMutableDictionary new];
-
-  Branch *branch = [self getInstance];
-  [branch crossPlatformIdDataWithCompletion:^(BranchCrossPlatformID *cpid) {
-    CDVPluginResult* pluginResult = nil;
-    if (cpid) {
-      // Convert the ObjC object back into JSON.  Should have kept the raw JSON response.
-      [json setObject:cpid.developerID forKey:@"developer_identity"];
-      [json setObject:cpid.crossPlatformID forKey:@"cross_platform_id"];
-      [json setObject:cpid.pastCrossPlatformIDs forKey:@"past_cross_platform_ids"];
-
-      NSMutableArray *probCPIDs = [NSMutableArray new];
-      for (BranchProbabilisticCrossPlatformID *tmp in cpid.probabiliticCrossPlatformIDs) {
-        if (tmp.crossPlatformID && tmp.score) {
-          NSMutableDictionary *pair = [NSMutableDictionary new];
-          [pair setObject:tmp.crossPlatformID forKey:@"id"];
-          [pair setObject:tmp.score forKey:@"probability"];
-          [probCPIDs addObject:pair];
-        }
-      }
-      [json setObject:probCPIDs forKey:@"prob_cross_platform_ids"];
-
-      pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:json];
-    } else {
-      pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:@"No CPIDs available"];
-    }
-    [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
-  }];
-}
 
 - (void)lastAttributedTouchData:(CDVInvokedUrlCommand *)command {
   NSMutableDictionary *json = [NSMutableDictionary new];